server/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
SQLAlchemy==2.0.19
Werkzeug==2.3.6
PyJWT==2.7.0
python-dotenv==1.0.0
